Add files via upload
My attempt of a Looping Load Balancer. There are a list of requests, and a number of servers to process those requests. The requests are processed sequentialy for the servers, i.e. when a request is processed, the next request will be processed by the next server and so on. 
- When there are more requests than servers: If the last server in the sequence is used, the next request is processed by the first server in the list (looping through the servers)
- When there are more servers than requests: the number of servers used is equal to the number of requests
- When there are no servers available: each request will not be processed** (no output)